Move extra buildings submit login down to ExtraBuildingsField
